Move/rename contents of 'docs' subpackage to 'signatures' module, to disambiguate from Sphinx docs
