+++ aiohttp_client_cache/backends/base.py
@@ -11,8 +11,8 @@ from aiohttp.typedefs import StrOrURL
 
 from aiohttp_client_cache.cache_control import CacheActions, ExpirationPatterns, ExpirationTime
 from aiohttp_client_cache.cache_keys import create_key
-from aiohttp_client_cache.docs.forge_utils import extend_init_signature
 from aiohttp_client_cache.response import AnyResponse, CachedResponse
+from aiohttp_client_cache.signatures import extend_init_signature
 
 ResponseOrKey = Union[CachedResponse, bytes, str, None]
 logger = getLogger(__name__)
