+++ aiohttp_client_cache/backends/dynamodb.py
@@ -7,7 +7,7 @@ from aioboto3.session import Session as AWSSession
 from botocore.exceptions import ClientError
 
 from aiohttp_client_cache.backends import BaseCache, CacheBackend, ResponseOrKey, get_valid_kwargs
-from aiohttp_client_cache.docs import dynamodb_template, extend_init_signature
+from aiohttp_client_cache.signatures import dynamodb_template, extend_init_signature
 
 
 @extend_init_signature(CacheBackend, dynamodb_template)
