+++ aiohttp_client_cache/backends/mongodb.py
@@ -3,7 +3,7 @@ from typing import AsyncIterable
 from motor.motor_asyncio import AsyncIOMotorClient
 
 from aiohttp_client_cache.backends import BaseCache, CacheBackend, ResponseOrKey, get_valid_kwargs
-from aiohttp_client_cache.docs import extend_init_signature, mongo_template
+from aiohttp_client_cache.signatures import extend_init_signature, mongo_template
 
 
 @extend_init_signature(CacheBackend, mongo_template)
