+++ aiohttp_client_cache/backends/sqlite.py
@@ -10,7 +10,7 @@ from typing import AsyncIterable, AsyncIterator, Union
 import aiosqlite
 
 from aiohttp_client_cache.backends import BaseCache, CacheBackend, ResponseOrKey, get_valid_kwargs
-from aiohttp_client_cache.docs import extend_init_signature, sqlite_template
+from aiohttp_client_cache.signatures import extend_init_signature, sqlite_template
 
 
 @extend_init_signature(CacheBackend, sqlite_template)
