--- aiohttp_client_cache/docs/__init__.py
@@ -1,3 +1,0 @@
-# flake8: noqa: F401, F403
-from aiohttp_client_cache.docs.connections import *
-from aiohttp_client_cache.docs.forge_utils import *
