+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "aiohttp-client-cache"
-version = "0.7.3"
+version = "0.7.4"
 description = "Persistent cache for aiohttp requests"
 authors = ["Jordan Cook"]
 license = "MIT License"
